Add input style, link style, profile link style
Added some pizzazz to the boring input because, let's face it, who wants boring? ¯\_(ツ)_/¯

Decided to transform the confusing link tag into a vibrant shade of green and threw in an outrageous amount of "very" to make it super-duper, incredibly, mind-bogglingly cool when hovered over. You won't believe your eyes!

Oh, and I went ahead and added some unnecessary style for profile linking in the "How To Use" section, just in case my intoxicated brain thought it was a good idea. My apologies if that came across as rude, but hey, blame the brain, right?

Thanks for sticking around until the end. It was a wild ride of wasted time and absolutely nothing else. And hey, I never asked you to read all this, but I appreciate it anyway. Cheers! 🍻